Applied pre-release C# style cleanup across the codebase.
    -- Wrapped 12 over-length lines, tightened per-method param docs, and rewrote ConfigLoader exception tags from "Thrown when X" to direct condition clauses.
    -- Standardized catch (Exception ex) to catch (Exception exception) at the three remaining sites in Task and MQTTClient.
    -- Renamed Display.cs to DisplayType.cs, Monitor.EnumeratedMonitors to EnumerateMonitors, and LayoutSettings.SubBox/MainBox.style to Style, updating every caller.
    -- Renamed Camera cam locals, Collider parameter shadows, and the task GameObject local to non-shadowing identifiers.
    -- Hoisted four scattered constants in CreateTask to the constants block and extracted a RandomSeedSentinel constant in Task.
    -- Repositioned StimulusTriggerZone.IsOccupancyMode to follow its field grouping and cached the MeshRenderer in Start instead of re-resolving it per call.
    -- Consolidated the two divergent MenuSettings<T> nested classes into a shared Gimbl/Editor/MenuSettings.cs and migrated DisplaysWindow off the camelCase 'selected' field.
    -- Guarded Monitor, FullScreenView, FullScreenViewManager, and FullScreenViewsSaved with #if UNITY_EDITOR so non-editor builds compile.
    -- Made TagsAndLayers a static class and inlined MQTTChannel.Init into the constructor, marking topic and client readonly.
    -- Fixed LinearTreadmill MQTT listener leak by storing the channel and unsubscribing it in OnDestroy.
    -- Fixed MiniJson culture invariance for numeric serialization and added defensive parsing for true/false/null literals.
    -- Removed dead MainWindow._window field, the dead "Paths" switch case, the unused CreateFunc<T> delegate, and the partial keyword from ActorObject.
    -- Changed OccupancyZone.GetElapsedMilliseconds return type from float to long; extracted the "_brightness" string in PerspectiveProjection to a named constant; reordered UpdateView to null-check projectionScreen first.

+++ Assets/InfiniteCorridorTask/Scripts/OccupancyGuidanceZone.cs
@@ -55,7 +55,7 @@ namespace SL.Tasks
         }
 
         /// <summary>Sets the zone state to active when the animal enters the guidance zone collider.</summary>
-        /// <param name="other">The collider that entered or exited the trigger zone.</param>
+        /// <param name="other">The collider that entered the trigger zone.</param>
         private void OnTriggerEnter(Collider other)
         {
             inZone = true;
@@ -69,7 +69,7 @@ namespace SL.Tasks
         }
 
         /// <summary>Sets the zone state to inactive when the animal exits the guidance zone collider.</summary>
-        /// <param name="other">The collider that entered or exited the trigger zone.</param>
+        /// <param name="other">The collider that exited the trigger zone.</param>
         private void OnTriggerExit(Collider other)
         {
             inZone = false;
@@ -87,9 +87,9 @@ namespace SL.Tasks
         private void TriggerBrakeActivation()
         {
             // Calculates remaining duration based on how much time has already elapsed.
-            float elapsedMilliseconds = _parentOccupancyZone.GetElapsedMilliseconds();
+            long elapsedMilliseconds = _parentOccupancyZone.GetElapsedMilliseconds();
             uint remainingMilliseconds = (uint)
-                Mathf.Max(0, _parentOccupancyZone.occupancyDurationMs - elapsedMilliseconds);
+                Mathf.Max(0f, _parentOccupancyZone.occupancyDurationMs - elapsedMilliseconds);
 
             Debug.Log($"OccupancyGuidanceZone: Triggering brake for {remainingMilliseconds}ms.");
