+++ Assets/InfiniteCorridorTask/Scripts/OccupancyZone.cs
@@ -63,7 +63,7 @@ namespace SL.Tasks
         }
 
         /// <summary>Starts the occupancy timer when the animal enters the zone collider.</summary>
-        /// <param name="other">The collider that entered or exited the trigger zone.</param>
+        /// <param name="other">The collider that entered the trigger zone.</param>
         private void OnTriggerEnter(Collider other)
         {
             if (!isActive || boundaryDisarmed)
@@ -75,7 +75,7 @@ namespace SL.Tasks
         }
 
         /// <summary>Stops the timer and checks the result when the animal exits the zone collider.</summary>
-        /// <param name="other">The collider that entered or exited the trigger zone.</param>
+        /// <param name="other">The collider that exited the trigger zone.</param>
         private void OnTriggerExit(Collider other)
         {
             if (!isActive)
@@ -91,7 +91,7 @@ namespace SL.Tasks
         }
 
         /// <summary>Returns the elapsed time in milliseconds since the occupancy timer started.</summary>
-        public float GetElapsedMilliseconds()
+        public long GetElapsedMilliseconds()
         {
             return _occupancyTimer.ElapsedMilliseconds;
         }
